change all color => colour, reorganise files, create custom hook
