+++ frontend/src/hooks/usePlayerColour.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+import { PlayerContext } from "../contexts/Context";
+
+export const usePlayerColour = () => {
+  const { playerColour } = useContext(PlayerContext)
+  return playerColour
+}
